+++ commfuncs.h
@@ -28,7 +28,7 @@
 
 using namespace std;
 
-#define VERSION "v0.99"
+#define VERSION "v0.991"
 
 #define UNKNOWN 0